Suppress console log output during build.
Recent changes to fix issue #4 added some actual log statements to make
sure that the compilation of log messages are happening correctly.
This slightly cleans up that code and adds a "quiet" logger in Logback
so that it doesn't create a bunch of output (which makes it look like
something is wrong).

Fixes #6

Still left to do is issue #7, which will take these thread of
development even further by closing the loop in tests.