Fix `getLogger` for higher-kinded types
Previously a top-level class declaration would fail on `getLogger` if
the class had a higher-kinded type parameter (e.g., `class A[F[_]]`).
Scala doesn't have a trivial way to say "I don't care about the
parameter" in this case, as we can with `WildcardType` in other cases.
So we simply fall back to getting the logger by name in this case.