Add wrappers for logged events in test appender
This adds a Scala-like wrapper around the events returned by
`TestAppender`. These are `AnyVal` instances, so the original event is
still available: it's all compiler magic. There's a few things that
aren't available through this wrapper, and there's a little overhead on
certain methods that return an `Option` rather than nullable values, but
this shouldn't be a big deal.