Add MiMa to the build
Log4s is a cross-cutting library where it's important to maintain binary
compatibility to stay out of jar hell. API changes have been almost
completely nonexistent up to this point so it hasn't been an issue, but
the upcoming Scala.JS support is a much more substantial change.

This also modifies the `release` command to check binary compatibility
for all versionas part of the release process.